+++ bin-proto-derive/src/lib.rs
@@ -69,11 +69,12 @@ fn impl_for_struct(
             let (reads, initializers) = codegen::reads(&strukt.fields);
             (
                 quote!(
-                    #[allow(unused_variables)]
-                    fn read(__io_reader: &mut dyn ::bin_proto::BitRead,
-                            __byte_order: ::bin_proto::ByteOrder,
-                            __ctx: &mut #ctx_ty)
-                            -> ::bin_proto::Result<Self> {
+                    fn read(
+                        __io_reader: &mut dyn ::bin_proto::BitRead,
+                        __byte_order: ::bin_proto::ByteOrder,
+                        __ctx: &mut #ctx_ty,
+                        __tag: (),
+                    ) -> ::bin_proto::Result<Self> {
                         #reads
                         Ok(Self #initializers)
                     }
@@ -85,11 +86,10 @@ fn impl_for_struct(
             let writes = codegen::writes(&strukt.fields, true);
             (
                 quote!(
-                    #[allow(unused_variables)]
                     fn write(&self, __io_writer: &mut dyn ::bin_proto::BitWrite,
-                             __byte_order: ::bin_proto::ByteOrder,
-                             __ctx: &mut #ctx_ty)
-                             -> ::bin_proto::Result<()> {
+                        __byte_order: ::bin_proto::ByteOrder,
+                        __ctx: &mut #ctx_ty
+                    ) -> ::bin_proto::Result<()> {
                         #writes
                         Ok(())
                     }
@@ -122,12 +122,12 @@ fn impl_for_enum(
         Operation::Read => {
             let read_variant = codegen::enums::read_variant_fields(&plan);
             let impl_body = quote!(
-                #[allow(unused_variables)]
-                fn read(__io_reader: &mut dyn ::bin_proto::BitRead,
-                        __byte_order: ::bin_proto::ByteOrder,
-                        __ctx: &mut #ctx_ty,
-                        __tag: __Tag)
-                        -> ::bin_proto::Result<Self> {
+                fn read(
+                    __io_reader: &mut dyn ::bin_proto::BitRead,
+                    __byte_order: ::bin_proto::ByteOrder,
+                    __ctx: &mut #ctx_ty,
+                    __tag: (__Tag,)
+                ) -> ::bin_proto::Result<Self> {
                     Ok(#read_variant)
                 }
             );
@@ -139,13 +139,14 @@ fn impl_for_enum(
 
             let read_discriminant = read_discriminant(&attribs);
             let impl_body = quote!(
-                #[allow(unused_variables)]
-                fn read(__io_reader: &mut dyn ::bin_proto::BitRead,
-                        __byte_order: ::bin_proto::ByteOrder,
-                        __ctx: &mut #ctx_ty)
-                        -> ::bin_proto::Result<Self> {
+                fn read(
+                    __io_reader: &mut dyn ::bin_proto::BitRead,
+                    __byte_order: ::bin_proto::ByteOrder,
+                    __ctx: &mut #ctx_ty,
+                    __tag: (),
+                ) -> ::bin_proto::Result<Self> {
                     let __tag: #discriminant_ty = #read_discriminant?;
-                    <Self as ::bin_proto::TaggedRead<_, _>>::read(__io_reader, __byte_order, __ctx, __tag)
+                    <Self as ::bin_proto::ProtocolRead<_, (#discriminant_ty,)>>::read(__io_reader, __byte_order, __ctx, (__tag,))
                 }
             );
             let protocol_read_impl = impl_trait_for(ast, &impl_body, &TraitImplType::ProtocolRead);
@@ -158,7 +159,6 @@ fn impl_for_enum(
         Operation::Write => {
             let write_variant = codegen::enums::write_variant_fields(&plan);
             let impl_body = quote!(
-                #[allow(unused_variables)]
                 fn write(&self,
                          __io_writer: &mut dyn ::bin_proto::BitWrite,
                          __byte_order: ::bin_proto::ByteOrder,
@@ -169,13 +169,12 @@ fn impl_for_enum(
                 }
             );
             let externally_tagged_write_impl =
-                impl_trait_for(ast, &impl_body, &TraitImplType::UntaggedWrite);
+                impl_trait_for(ast, &impl_body, &TraitImplType::TaggedWrite);
 
             let variant_discriminant = variant_discriminant(&plan, &attribs);
             let impl_body = quote!(
                 type Discriminant = #discriminant_ty;
 
-                #[allow(unused_variables)]
                 fn discriminant(&self) -> Self::Discriminant {
                     #variant_discriminant
                 }
@@ -184,14 +183,13 @@ fn impl_for_enum(
 
             let write_discriminant = write_discriminant(&attribs);
             let impl_body = quote!(
-                #[allow(unused_variables)]
                 fn write(&self,
                          __io_writer: &mut dyn ::bin_proto::BitWrite,
                          __byte_order: ::bin_proto::ByteOrder,
                          __ctx: &mut #ctx_ty)
                          -> ::bin_proto::Result<()> {
                     #write_discriminant
-                    <Self as ::bin_proto::UntaggedWrite<_>>::write(self, __io_writer, __byte_order, __ctx)
+                    <Self as ::bin_proto::ProtocolWrite<_, ::bin_proto::Untagged>>::write(self, __io_writer, __byte_order, __ctx)
                 }
             );
             let protocol_write_impl =