+++ bin-proto/src/impls/cstring.rs
@@ -3,10 +3,10 @@ use alloc::{ffi::CString, vec::Vec};
 use crate::{util, BitRead, BitWrite, ByteOrder, ProtocolRead, ProtocolWrite, Result};
 
 impl<Ctx> ProtocolRead<Ctx> for CString {
-    fn read(read: &mut dyn BitRead, byte_order: ByteOrder, ctx: &mut Ctx) -> Result<Self> {
+    fn read(read: &mut dyn BitRead, byte_order: ByteOrder, ctx: &mut Ctx, tag: ()) -> Result<Self> {
         let mut result = Vec::new();
         loop {
-            let c: u8 = ProtocolRead::read(read, byte_order, ctx)?;
+            let c: u8 = ProtocolRead::read(read, byte_order, ctx, tag)?;
             if c == 0x00 {
                 return Ok(Self::new(result)?);
             }