+++ bin-proto/src/impls/list.rs
@@ -3,7 +3,7 @@ macro_rules! impl_read_list {
         $ty:ident<T $(: $tbound0:ident $(+ $tbound1:ident)?)?
         $(, $h:ident: $hbound0:ident + $hbound1:ident)?>
     ) => {
-        impl<Tag, Ctx, T, $($h)?> $crate::TaggedRead<Tag, Ctx> for $ty<T, $($h)?>
+        impl<Tag, Ctx, T, $($h)?> $crate::ProtocolRead<Ctx, (Tag,)> for $ty<T, $($h)?>
         where
             T: $crate::ProtocolRead<Ctx> $(+ $tbound0 $(+ $tbound1)?)?,
             Tag: ::core::convert::TryInto<usize>,
@@ -12,10 +12,10 @@ macro_rules! impl_read_list {
             fn read(read: &mut dyn $crate::BitRead,
                 byte_order: $crate::ByteOrder,
                 ctx: &mut Ctx,
-                tag: Tag,
+                tag: (Tag,),
             ) -> $crate::Result<Self> {
                 let elements = $crate::util::read_items(
-                    ::core::convert::TryInto::try_into(tag)
+                    ::core::convert::TryInto::try_into(tag.0)
                         .map_err(|_| $crate::Error::TagConvert)?,
                     read,
                     byte_order,
@@ -26,7 +26,7 @@ macro_rules! impl_read_list {
         }
 
 
-        impl<Ctx, T, $($h)?> $crate::FlexibleArrayMemberRead<Ctx> for $ty<T, $($h)?>
+        impl<Ctx, T, $($h)?> $crate::ProtocolRead<Ctx, $crate::Untagged> for $ty<T, $($h)?>
         where
             T: $crate::ProtocolRead<Ctx> $(+ $tbound0 $(+ $tbound1)?)?,
             $($h: $hbound0 + $hbound1)?
@@ -34,7 +34,8 @@ macro_rules! impl_read_list {
             fn read(
                 read: &mut dyn $crate::BitRead,
                 byte_order: $crate::ByteOrder,
-                ctx: &mut Ctx
+                ctx: &mut Ctx,
+                _: $crate::Untagged,
             ) -> $crate::Result<Self> {
                 Ok(::core::iter::IntoIterator::into_iter(
                     $crate::util::read_items_to_eof(read, byte_order, ctx)?
@@ -46,7 +47,7 @@ macro_rules! impl_read_list {
 
 macro_rules! impl_write_list {
     ( $ty:ident<T $(: $tbound0:ident $(+ $tbound1:ident)?)? $(, $h:ident)?> ) => {
-        impl<Ctx, T, $($h)?> $crate::UntaggedWrite<Ctx> for $ty<T, $($h)?>
+        impl<Ctx, T, $($h)?> $crate::ProtocolWrite<Ctx, $crate::Untagged> for $ty<T, $($h)?>
         where
             T: $crate::ProtocolWrite<Ctx> $(+ $tbound0 $(+ $tbound1)?)?
         {
@@ -66,7 +67,7 @@ mod vec {
 
     impl_read_list!(Vec<T>);
     impl_write_list!(Vec<T>);
-    test_flexible_array_member_read_and_tagged!(Vec<u8>| 3: alloc::vec![1, 2, 3] => [0x01, 0x02, 0x03]);
+    test_flexible_array_member_read_and_protocol!(Vec<u8>| 3: alloc::vec![1, 2, 3] => [0x01, 0x02, 0x03]);
 }
 
 mod linked_list {
@@ -74,7 +75,7 @@ mod linked_list {
 
     impl_read_list!(LinkedList<T>);
     impl_write_list!(LinkedList<T>);
-    test_flexible_array_member_read_and_tagged!(LinkedList<u8>| 3: [1, 2, 3].into() => [0x01, 0x02, 0x03]);
+    test_flexible_array_member_read_and_protocol!(LinkedList<u8>| 3: [1, 2, 3].into() => [0x01, 0x02, 0x03]);
 }
 
 mod vec_deque {
@@ -82,7 +83,7 @@ mod vec_deque {
 
     impl_read_list!(VecDeque<T>);
     impl_write_list!(VecDeque<T>);
-    test_flexible_array_member_read_and_tagged!(VecDeque<u8>| 3: [1, 2, 3].into() => [0x01, 0x02, 0x03]);
+    test_flexible_array_member_read_and_protocol!(VecDeque<u8>| 3: [1, 2, 3].into() => [0x01, 0x02, 0x03]);
 }
 
 mod b_tree_set {
@@ -90,7 +91,7 @@ mod b_tree_set {
 
     impl_read_list!(BTreeSet<T: Ord>);
     impl_write_list!(BTreeSet<T: Ord>);
-    test_flexible_array_member_read_and_tagged!(BTreeSet<u8>| 3: [1, 2, 3].into() => [0x01, 0x02, 0x03]);
+    test_flexible_array_member_read_and_protocol!(BTreeSet<u8>| 3: [1, 2, 3].into() => [0x01, 0x02, 0x03]);
 }
 
 mod binary_heap {
@@ -107,5 +108,5 @@ mod hash_set {
 
     impl_read_list!(HashSet<T: Hash + Eq, H: BuildHasher + Default>);
     impl_write_list!(HashSet<T: Hash + Eq, H>);
-    test_flexible_array_member_read_and_tagged!(HashSet<u8>| 1: [1].into() => [0x01]);
+    test_flexible_array_member_read_and_protocol!(HashSet<u8>| 1: [1].into() => [0x01]);
 }