+++ bin-proto/src/impls/string.rs
@@ -1,11 +1,10 @@
 use crate::{
-    util, BitRead, BitWrite, ByteOrder, Error, FlexibleArrayMemberRead, Result, TaggedRead,
-    UntaggedWrite,
+    util, BitRead, BitWrite, ByteOrder, Error, ProtocolRead, ProtocolWrite, Result, Untagged,
 };
 
 use alloc::{string::String, vec::Vec};
 
-impl<Tag, Ctx> TaggedRead<Tag, Ctx> for String
+impl<Tag, Ctx> ProtocolRead<Ctx, (Tag,)> for String
 where
     Tag: TryInto<usize>,
 {
@@ -13,10 +12,10 @@ where
         read: &mut dyn BitRead,
         byte_order: ByteOrder,
         ctx: &mut Ctx,
-        tag: Tag,
+        tag: (Tag,),
     ) -> Result<Self> {
         let bytes = util::read_items(
-            tag.try_into().map_err(|_| Error::TagConvert)?,
+            tag.0.try_into().map_err(|_| Error::TagConvert)?,
             read,
             byte_order,
             ctx,
@@ -25,18 +24,23 @@ where
     }
 }
 
-impl<Ctx> UntaggedWrite<Ctx> for String {
+impl<Ctx> ProtocolWrite<Ctx, Untagged> for String {
     fn write(&self, write: &mut dyn BitWrite, byte_order: ByteOrder, ctx: &mut Ctx) -> Result<()> {
         let bytes: Vec<_> = self.bytes().collect();
         util::write_items(&bytes, write, byte_order, ctx)
     }
 }
 
-impl<Ctx> FlexibleArrayMemberRead<Ctx> for String {
-    fn read(read: &mut dyn BitRead, byte_order: ByteOrder, ctx: &mut Ctx) -> Result<Self> {
+impl<Ctx> ProtocolRead<Ctx, Untagged> for String {
+    fn read(
+        read: &mut dyn BitRead,
+        byte_order: ByteOrder,
+        ctx: &mut Ctx,
+        _: Untagged,
+    ) -> Result<Self> {
         let bytes = util::read_items_to_eof(read, byte_order, ctx)?;
         Ok(Self::from_utf8(bytes)?)
     }
 }
 
-test_flexible_array_member_read_and_tagged!(String| 3: "abc".into() => [b'a', b'b', b'c']);
+test_flexible_array_member_read_and_protocol!(String| 3: "abc".into() => [b'a', b'b', b'c']);